Major update (Frontend)
Finished a very very quick frontend for the website. It really looks horrible but gets the job done.
Fixed bugs along the way too. Now the only things to do are to add the edit and delete buttons, add a readme file, fix
all files according to PEP8 and finally write a database populator script.